Allow user-defined local MathJax installation.
Allow user-defined directory for a local MathJax installation on the
server. The user may set the variable `$wgLocalMathJaxDir` to a
directory relative to server root that contains the MathJax
javascripts. For example, if the directory `MathJax-2.7.1` contains
the `MathJax.js` script, one would add to their `LocalSettings.php`
the following line:
`wgLocalMathJaxDir = '/MathJax-2.7.1';`
and their mediawiki instance should now render math using MathJax
installed on their server rather than on a CDN. **Note** that the
leading '/' is important!